ai-hugo-backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.1.1
python-dotenv==1.0.0
httpx==0.24.1
aiofiles==23.1.0
aiohttp==3.9.3
pyyaml==6.0.1
